activityPrice crud done

+++ src/script/database.sql
@@ -89,7 +89,7 @@ create table activityPhoto(
 create table activityPrice(
     priceID  serial primary key,
     activityID varchar(10) not null references activity(activityID),
-    hourPrice NUMERIC(15, 2) not null DEFAULT 0,
+    hourPrice INTEGER not null DEFAULT 0,
     price NUMERIC(15, 2) not null DEFAULT 0,
     dateChanged date default current_date,
     status integer check (status >= 0)
